Detect SVN revision and path on Git working copy
Currently when building from Git `clang --version` shows Git hash and repo URL
instead of SVN revision. However SVN is master for LLVM, so it is better to
show exactly the same as if built from SVN, if we build using clean master.

Therefore we check if last commit contains git-svn-id: and if so we extract SVN
revision and repo URL. Otherwise we fallback to Git hash and URL.

Calls to sed and grep conform to POSIX standard (no extended RE).